requirements.txt
# HuggingFace Papers Aggregation System
# Python 3.10+

# HTTP & Async
aiohttp>=3.9.0
httpx>=0.25.0
aiofiles>=23.2.1

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Data Processing
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Notion API
notion-client>=2.2.0

# Logging
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# Date handling
python-dateutil>=2.8.0
pysocks
requests
pyyaml
aiohttp

# Optional: for development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0